+++ src/main/kotlin/io/jterrier/wheels/views/utils.kt
@@ -1,3 +1,3 @@
 package io.jterrier.wheels.views
 
-fun Float.toKm(): Double = this.toDouble() / 1000
+fun Double.toKm(): Double = this / 1000
